Further clean up eBOSSLens()
