+++ eBOSSLens.py
@@ -13,7 +13,10 @@ from SDSSObject import SDSSObject
 from objFilter import qsoFilter, genFilter
 from peakFinder import bolEstm, peakCandidate, combNear, checkFore, \
     qsoContfit, qsoBggal, jpLens, doubletO2, skewFit
-from resultSave import galSave, plotGalaxyLens
+from galSave import galSave, plotGalaxyLens
+from qsoSave import qsoSave
+from jptSave import jptSave
+from lyaSave import lyaSave
 from utils import gauss
 
 
@@ -129,7 +132,6 @@ def eBOSSLens(plate, mjd, fiberid, searchLyA, QSOlens, Jackpot, max_chi2=4.0,
             peak.wavelength = params[0]
             peak[2] = params[1]
             peak[3] = params[2]
-            #eq_Width = quad(gauss,x0-200,x0+200,args=(params[0],params[1],params[2]))
             chi2_width = chisq
             peak[15] = chisq
         '''
@@ -169,7 +171,7 @@ def eBOSSLens(plate, mjd, fiberid, searchLyA, QSOlens, Jackpot, max_chi2=4.0,
                 break
     # TODO: clean up
     '''
-    elif searchLyA == False and QSOlens == True and Jackpot == False:
+    elif (not (searchLyA or Jackpot)) and QSOlens:
         peak_candidates = n.array([peak for peak in peak_candidates if peak[5]>(1.5+peak[6])])
         if len(peak_candidates) == 0:
             continue
@@ -195,122 +197,11 @@ def eBOSSLens(plate, mjd, fiberid, searchLyA, QSOlens, Jackpot, max_chi2=4.0,
     if not below_9500:
         raise Exception("Rejected since no below 9200")
     # Try to infer background redshift
-    detection = False
     if not (searchLyA or QSOlens or Jackpot):
         detection = galSave(doublet, obj, peak_candidates, doublet_index,
                             savedir, em_lines)
-    # TODO: clean up
-    '''
-    elif searchLyA == False and QSOlens == True and Jackpot==False:
-        for k in range(len(peak_candidates)):
-            z_backgal = peak_candidates[k][4]
-            peak = peak_candidates[k]
-            # compute OII,OIII flux (of lensed galaxy)
-            temp_fluxes_OII = n.zeros(5)
-            temp_fluxes_OIII = n.zeros(5)
-            dwave = n.array([wave[gen_i+1]-wave[gen_i] if gen_i<len(wave)-1 else 0 for gen_i in range(len(wave))])
-            if z_backgal < 1:
-                for j in range(4,9):
-                    temp_bounds = n.linspace(wave2bin((1+z_backgal)*3727,c0,c1,Nmax)-j,wave2bin((1+z_backgal)*3727,c0,c1,Nmax)+j,2*j+1,dtype = n.int16)
-                    temp_fluxes_OII[j-4] = n.sum((flux[i,temp_bounds]-synflux[i,temp_bounds])*dwave[temp_bounds])
-                    temp_bounds = n.linspace(wave2bin((1+z_backgal)*5007,c0,c1,Nmax)-j,wave2bin((1+z_backgal)*5007,c0,c1,Nmax)+j,2*j+1,dtype = n.int16)
-                    temp_fluxes_OIII[j-4] = n.sum((flux[i,temp_bounds]-synflux[i,temp_bounds])*dwave[temp_bounds])
-            OII_flux = n.median(temp_fluxes_OII)
-            OIII_flux = n.median(temp_fluxes_OIII)
-            fileQSO = open(savedir +  '/candidates_QSO.txt','a')
-            fileQSO.write('\n' + str([RA[i], DEC[i], int(plate), int(mjd), fiberid[i], z[i], peak[0],peak[4],peak[5],peak[6],spectroflux[i,1], spectroflux[i,3], OII_flux, OIII_flux ]))
-            fileQSO.close()
-            plot_QSOGal(k=k,RA = RA[i],DEC= DEC[i],plate = int(plate), mjd = int(mjd), fiberid = fiberid[i],z=z[i], z_backgal= z_backgal,flux=flux[i,:],wave=wave,synflux=synflux[i,:],ivar= ivar[i,:], \
-                reduced_flux = reduced_flux[i,:], c0=c0,c1=c1,Nmax=Nmax,show = plot_show,topdir=topdir, savedir=savedir, HB_wave = HB_wave , params_beta=params_beta, line_coeff =line_coeff)
-            # plot with +-1 AA for paper
-            if paper:
-                plot_QSOGal(k=k+len(peak_candidates),RA = RA[i],DEC= DEC[i],plate = int(plate), mjd = int(mjd), fiberid = fiberid[i],z=z[i], z_backgal= z_backgal+0.0005,flux=flux[i,:],wave=wave,synflux=synflux[i,:],ivar= ivar[i,:], \
-                reduced_flux = reduced_flux[i,:], c0=c0,c1=c1,Nmax=Nmax,show = plot_show,topdir=topdir, savedir=savedir, HB_wave = HB_wave , params_beta=params_beta, line_coeff =line_coeff)
-                plot_QSOGal(k=k+len(peak_candidates)+1,RA = RA[i],DEC= DEC[i],plate = int(plate), mjd = int(mjd), fiberid = fiberid[i],z=z[i], z_backgal= z_backgal-0.0005,flux=flux[i,:],wave=wave,synflux=synflux[i,:],ivar= ivar[i,:], \
-                reduced_flux = reduced_flux[i,:], c0=c0,c1=c1,Nmax=Nmax,show = plot_show,topdir=topdir, savedir=savedir, HB_wave = HB_wave , params_beta=params_beta, line_coeff =line_coeff)
-    elif Jackpot == True:
-        k = 0
-        for peak in peak_candidates:
-            fileJ = open(topdir + savedir +  '/candidates_Jackpot.txt','a')
-            fileJ.write('\n' + str([RA[i], DEC[i], int(plate), int(mjd), fiberid[i], z[i], peak[2],peak[3],peak[4],peak[5],peak[6],spectroflux[i,1], spectroflux[i,3]]))
-            fileJ.close()
-            k += 1
-            plot_Jackpot(RA= RA[i],DEC=DEC[i],plate =int(plate), mjd=int(mjd), fiberid=fiberid[i], z=z[i],wave=wave, flux =flux[i,:], synflux = synflux[i,:],topdir=topdir,savedir=savedir ,peak = peak, show = plot_show, counter = k)
-    elif searchLyA==True and QSOlens==True and Jackpot == False:
-        if QSOlens:
-            fileLyA = open(topdir + savedir +  '/candidates_QSO_LyA.txt','a')
-        else:
-            fileLyA = open(topdir + savedir +  '/candidates_LAE.txt','a')
-        n_peak = 1
-        for peak in peak_candidates:
-            x0 = peak[0]
-            z_O2 = x0/3727.24 - 1.0
-            # compute SN of Ha, Hb, OIII to conclude if LyA candidate is or not OII at low-redshift
-            SNlines = 0
-            em_lines = n.array([4861.325,5006.843,6562.801])
-            for l in em_lines:
-                center_bin = wave2bin(l*(1+z_O2),c0,c1,Nmax)
-                SNlines += max(SN[center_bin-2:center_bin+2])**2
-            SNlines = n.sqrt(SNlines)
-            if SNlines < 100:
-                if n.abs(peak[9]) > n.abs(peak[13]):
-                    skewness = peak[9]
-                else:
-                    skewness = peak[13]
-                params = peak[1:6]
-                params_skew = peak[7:15]
-                if not(n.any(params) and n.any(params_skew)):
-                    continue
-                bounds = n.linspace(wave2bin(x0,c0,c1,Nmax)-15,wave2bin(x0,c0,c1,Nmax)+15,31,dtype = n.int16)
-                #compute equivalent width before manipulating the spectra
-                dwave = n.array([wave[gen_i+1]-wave[gen_i] if gen_i<len(wave)-1 else 0 for gen_i in range(len(wave))])
-                eq_Width = n.sum((flux[i,bounds]/synflux[i,bounds]-1)*dwave[bounds])
-                # compute LyA flux
-                temp_fluxes = n.zeros(5)
-                for j in range(4,9):
-                    temp_bounds = n.linspace(wave2bin(x0,c0,c1,Nmax)-j,wave2bin(x0,c0,c1,Nmax)+j,2*j+1,dtype = n.int16)
-                    temp_fluxes[j-4] = n.sum((flux[i,temp_bounds]-synflux[i,temp_bounds])*dwave[temp_bounds])
-                lyA_flux = n.median(temp_fluxes)
-                #compute skewness indicator (on reduced flux but without 3rd order fit (which is meant for bad fit QSO and not present in final candidates)
-                I = n.sum(reduced_flux[i,bounds])
-                xmean = n.sum(reduced_flux[i,bounds]*wave[bounds])/I
-                sigma2 = n.sum(reduced_flux[i,bounds]*(wave[bounds] - xmean)**2)/I
-                S = n.sum(reduced_flux[i,bounds]*(wave[bounds] - xmean)**3)/(I*n.sign(sigma2)*n.abs(sigma2)**1.5)
-                local_wave = wave[bounds]
-                local_flux = reduced_flux[i,bounds]
-                peak_index = local_flux.argmax()
-                F = 0.1*local_flux[peak_index]
-                #Find blue and red 10% peak flux
-                f = 10*F
-                k = peak_index
-                while f > F and 0<k:
-                    k = k-1
-                    f = local_flux[k]
-                a = (local_flux[k+1]-local_flux[k])/(local_wave[k+1]-local_wave[k])
-                b = local_flux[k] - a*local_wave[k]
-                l_blue_10 = (F-b)/a
-                k = peak_index
-                f = 10*F
-                while f > F and k<len(local_flux)-1:
-                    k = k+1
-                    f = local_flux[k]
-                a = (local_flux[k]-local_flux[k-1])/(local_wave[k]-local_wave[k-1])
-                b = local_flux[k-1] - a*local_wave[k-1]
-                l_red_10 = (F-b)/a
-                Sw = S*(l_red_10-l_blue_10)
-                a_lambda = (l_red_10-local_wave[peak_index])/(local_wave[peak_index]-l_blue_10)
-                # save the parameters
-                fileLyA.write('\n' + str([peak[0],peak[6],z[i], SNlines ,RA[i], DEC[i], int(plate), int(mjd), fiberid[i],params[0],params[1],params[2],params[3],params[4],
-                                params_skew[0],params_skew[1],params_skew[2],params_skew[3],params_skew[4],params_skew[5],params_skew[6],params_skew[7],peak[15],peak[16],peak[17],peak[18], skewness, S, Sw, a_lambda,rchi2[i],peak[19],spectroflux[i,1], spectroflux[i,3],lyA_flux]))
-                # Make the graph
-                plot_QSOLAE(RA= RA[i],DEC = DEC[i],z=z[i],flux=flux[i,:],wave=wave,synflux=synflux[i,:],x0= x0, ivar = ivar[i,:], reduced_flux = reduced_flux[i,:],window=window,peak =peak,
-                    params = params,params_skew=params_skew, topdir = topdir, savedir = savedir, n_peak = n_peak, plate = int(plate), mjd = int(mjd), fiberid = fiberid[i],c0=c0,c1=c1,Nmax=Nmax, show = plot_show, paper = paper, QSOlens = QSOlens)
-            n_peak = n_peak +1
-        fileLyA.close()
-        '''
-    # Save surviving candidates (Galaxy-Galaxy case)
-    peaks = []
-    if not (searchLyA or QSOlens or Jackpot):
+        # Save surviving candidates (Galaxy-Galaxy case)
+        peaks = []
         for k in range(len(peak_candidates)):
             peak = peak_candidates[k]
             if k == doublet_index and doublet:
@@ -333,3 +224,12 @@ def eBOSSLens(plate, mjd, fiberid, searchLyA, QSOlens, Jackpot, max_chi2=4.0,
                                   var=peaks[k][2])
             plotGalaxyLens(doublet, obj, savedir, peak_candidates,
                            doublet_index, fit)
+    elif (not (searchLyA or Jackpot)) and QSOlens:
+        # TODO: complete the function
+        qsoSave(peak_candidates, savedir)
+    elif Jackpot:
+        # TODO: complete the function
+        jptSave()
+    elif searchLyA and QSOlens and (not Jackpot):
+        # TODO: complete the function
+        lyaSave()
