+++ utils_QSO.py
@@ -144,242 +144,3 @@ def QSO_compute_FWHM(sobj, l_width):
         line_coeff = None
 
     return FWHM, l_times_luminosity, HB_wave, params, line_coeff
-
-def plot_QSOLAE(RA,DEC,z,flux,wave,synflux,x0,ivar, reduced_flux,window,peak,params,params_skew, topdir, savedir, n_peak, plate, mjd, fiberid,c0,c1,Nmax, show = False, paper=True, QSOlens = True):
-
-    if show ==False:
-        mpl.use('Agg')
-    # Create and save graph
-    fontP = FontProperties()
-    fontP.set_size('medium')
-    plt.figure(figsize=(12,3))
-    plt.suptitle(SDSSname(RA,DEC)+'\n'+'RA='+str(RA)+', Dec='+str(DEC) +', $z_{QSO}='+'{:03.3}'.format(z)+ '$')
-    plt.ylabel('$f_{\lambda}\, (10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}$')
-
-    if paper:
-        gs = gridspec.GridSpec(1,3)
-
-        smoothed_flux = np.array([np.mean(flux[ii-2:ii+3]) for ii in range(len(flux)) if (ii>4 and ii<len(flux)-4)])
-
-        p1 = plt.subplot(gs[0,:2])
-        #p1.plot(wave,  flux, 'k', label = 'BOSS Flux', drawstyle='steps-mid')
-        p1.plot(wave[5:-4], smoothed_flux, 'k', label = 'eBOSS Flux', drawstyle='steps-mid')
-        p1.plot(wave, synflux, 'r', label = 'PCA fit', drawstyle='steps-mid')
-        p1.set_ylim(np.min(synflux)-3, np.max(synflux)+3)
-
-        p1.vlines(x = x0,ymin= -100,ymax= 100,colors= 'g',linestyles='dashed')
-        box = p1.get_position()
-        p1.set_position([box.x0,box.y0+0.06,box.width,box.height*0.85])
-        plt.ylabel('Flux [$10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}$]')
-        plt.xlabel('Observed wavelength [$\AA$]')
-        p2 = plt.subplot(gs[0,2:3])
-
-        p2.plot(wave,  flux, 'k', label = 'eBOSS Flux', drawstyle='steps-mid')
-        p2.plot(wave, synflux, 'r', label = 'PCA fit', drawstyle='steps-mid')
-
-        p2.set_ylim(np.min(flux[window]), np.max(flux[window])+0.5)
-        p2.legend(loc='upper right', bbox_to_anchor = (1.3,1.1), ncol = 1, prop=fontP)
-        box = p2.get_position()
-        p2.set_position([box.x0,box.y0+0.06,box.width*0.9,box.height*0.85])
-        x1 = int(x0/10.)*10
-        plt.xticks([x1-10,x1,x1+10,x1+20])
-        p2.set_xlim(x0-15,x0+25)
-        plt.xlabel('Observed wavelength [$\AA$]')
-
-    else:
-
-        gs = gridspec.GridSpec(2,2)
-        p1 = plt.subplot(gs[0,:2])
-        p1.plot(wave,  flux[:], 'k', label = 'BOSS Flux', drawstyle='steps-mid')
-        p1.plot(wave, synflux[:], 'r', label = 'PCA fit', drawstyle='steps-mid')
-        #p1.plot(wave,ivar[:]-15, 'g', label = 'var$^{-1}-15$')
-
-        p1.fill_between(wave,np.min(synflux[:])-10,np.max(synflux[:])+10,where=(ivar[:]==0),facecolor='k', alpha=0.2)
-        p1.fill_between(wave,np.min(synflux[:])-10,np.max(synflux[:])+10,where=np.logical_and(5560<wave, wave<5600),facecolor='c', alpha=0.2)
-        p1.fill_between(wave,np.min(synflux[:])-10,np.max(synflux[:])+10,where=np.logical_and(5880<wave, wave<5905),facecolor='c', alpha=0.2)
-        p1.fill_between(wave,np.min(synflux[:])-10,np.max(synflux[:])+10,where=np.logical_and(6285<wave, wave<6315),facecolor='c', alpha=0.2)
-        p1.fill_between(wave,np.min(synflux[:])-10,np.max(synflux[:])+10,where=np.logical_and(6348<wave,wave<6378),facecolor='c', alpha=0.2)
-
-        p1.legend(loc='upper right', bbox_to_anchor = (1,1), ncol = 1, prop=fontP)
-        box = p1.get_position()
-        p1.set_position([box.x0,box.y0,box.width,box.height])
-
-        p1.set_ylim(np.min(synflux[:])-3, np.max(synflux[:])+3)
-        plt.ylabel('$f_{\lambda}\, [10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}]$')
-        if QSOlens == False:
-            p1.set_xlim(3600,6000)
-
-
-        window = np.linspace(wave2bin(x0,c0,c1,Nmax)-40,wave2bin(x0,c0,c1,Nmax)+40,81,dtype = np.int16)
-        if QSOlens:
-            p3 = plt.subplot(gs[1,:1])
-            p3.plot(wave,  flux[:], 'k', label = 'BOSS Flux', drawstyle='steps-mid')
-            p3.plot(wave, synflux[:], 'r', label = 'PCA fit', drawstyle='steps-mid')
-            p3.set_xlim(np.min(wave[window]),np.max(wave[window]))
-            p3.set_ylim(np.min(synflux[window])-1, np.max(flux[window])+1)
-            box = p3.get_position()
-
-            p3.set_position([box.x0,box.y0,box.width,box.height])
-            plt.ylabel('$f_{\lambda}\, [10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}]$')
-            p3.legend(loc='upper right', bbox_to_anchor = (1,1), ncol = 1,prop=fontP)
-
-            p3.locator_params(axis='x',nbins=6)
-
-            median_local = np.median(reduced_flux[window])
-            fit_QSO = np.poly1d(np.polyfit(x=wave[window],y=reduced_flux[window],deg=3,w=(np.abs(reduced_flux[window]-median_local)<5)*np.sqrt(ivar[window])) )
-            p4 = plt.subplot(gs[1,1:2])
-            p4.plot(wave[window], fit_QSO(wave[window]), '-m',label = 'Order 3 fit')
-            box = p4.get_position()
-            p4.set_xlim(np.min(wave[window]),np.max(wave[window]))
-
-            p4.set_position([box.x0,box.y0,box.width,box.height])
-            p4.plot(wave[window], reduced_flux[window],'k', label = 'Reduced flux', drawstyle='steps-mid')
-            p4.legend(loc='upper right', bbox_to_anchor = (1,1), ncol = 1,prop=fontP)
-
-            p4.locator_params(axis='x',nbins=6)
-        else:
-            p3 = plt.subplot(gs[1,:2])
-            p3.plot(wave,  flux[:], 'k', label = 'BOSS Flux', drawstyle='steps-mid')
-            p3.plot(wave, synflux[:], 'r', label = 'PCA fit', drawstyle='steps-mid')
-            p3.legend(prop=fontP)
-            p3.set_xlim(peak[0]-50,peak[0]+60)
-            p3.set_ylim(np.min(synflux[bounds])-2, np.max(flux[bounds])+3)
-            plt.ylabel('$f_{\lambda}\, [10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}]$', fontsize=18)
-
-
-
-        ##### Old code to
-        #p2 = plt.subplot(gs[2,:2])
-        #if QSOlens:
-            #p2.plot(wave[window], reduced_flux[window]-fit_QSO(wave[window]),'k', label = 'Reduced flux', drawstyle='steps-mid')
-        #else:
-            #p2.plot(wave[window], reduced_flux[window],'k', label = 'Reduced flux')
-        #if 0.0<peak[16]<peak[15]:
-            #p2.plot(wave,gauss2(x=wave,x1=params[0],x2=params[1],A1=params[2],A2=params[3],var=params[4]),'g', label = r'$\chi_D^2 = $' + '{:.4}'.format(peak[16]))
-        #else:
-            #p2.plot(wave,gauss(x=wave, x_0=params[0], A=params[1], var=params[2]),'r', label = r'$\chi_G^2 = $' + '{:.4}'.format(peak[15]) )
-        #if 0.0<peak[17]<peak[18]:
-            #p2.plot(wave,skew(x=wave,A = params_skew[0], w=params_skew[1], a=params_skew[2], eps=params_skew[3]), 'b', label =r'$\chi_S^2 = $' + '{:.4}'.format(peak[17]))
-        #else:
-            #p2.plot(wave,skew2(x=wave,A1 = params_skew[0], w1=params_skew[1], a1=params_skew[2], eps1 = params_skew[3], A2 = params_skew[4], w2=params_skew[5], a2=params_skew[6], eps2=params_skew[7]), 'c',label= r'$\chi_{S2}^2 = $' + '{:.4}'.format(peak[18]))
-        #box = p2.get_position()
-        #p2.set_position([box.x0,box.y0,box.width*0.9,box.height])
-        #p2.legend(loc='upper right', bbox_to_anchor = (1.2,1), ncol = 1,prop=fontP)
-        #plt.xlabel('$Wavelength\, [\AA]$',fontsize = 18)
-        #p2.set_xlim(np.min(wave[window]),np.max(wave[window]))
-        #if QSOlens:
-            #p2.set_ylim(-1, np.max(reduced_flux[window]-fit_QSO(wave[window])+1))
-
-    make_sure_path_exists(topdir + savedir +'/plots/')
-    plt.savefig(topdir + savedir +'/plots/'+SDSSname(RA,DEC)+ '-' + str(plate) + '-' + str(mjd) + '-' + str(fiberid) + '-' + str(n_peak)+ '.eps', format = 'eps', dpi = 2000)
-
-    if show:
-        plt.show()
-    else:
-        plt.close()
-
-
-def plot_QSOGal(k,RA,DEC,plate, mjd, fiberid,z, z_backgal,flux,wave,synflux,ivar, reduced_flux,c0,c1,Nmax,topdir, savedir,show = False, HB_wave = 0.0, params_beta=[0,0,0], line_coeff = [0,0]):
-
-    em_lines = np.array([3726.5,4861.325,4958.911,5006.843,6562.801])
-
-    if show ==False:
-        mpl.use('Agg')
-    fontP = FontProperties()
-    fontP.set_size('medium')
-
-    plt.suptitle(SDSSname(RA,DEC)+'\n'+'RA='+str(RA)+', Dec='+str(DEC) +', $z_{QSO}='+'{:03.3}'.format(z)+ '$')
-
-
-    gs = gridspec.GridSpec(2,4)
-    p1 = plt.subplot(gs[0,:4])
-
-
-    smoothed_flux = np.array([np.mean(flux[ii-2:ii+3]) for ii in range(len(flux[:])) if (ii>4 and ii<len(flux[:])-4)])
-
-
-    p1.plot(wave[5:-4], smoothed_flux, 'k', label = 'BOSS Flux', drawstyle='steps-mid')
-    #p1.plot(wave,  flux[:], 'k', label = 'BOSS Flux')
-    p1.plot(wave, synflux[:], 'r', label = 'PCA fit')
-
-    if z<1 and show == True:
-
-        p1.plot(HB_wave, lorentz(HB_wave, params_beta[0],params_beta[1],params_beta[2]) + HB_wave*line_coeff[0] + line_coeff[1], '--g')
-    box = p1.get_position()
-
-    p1.set_position([box.x0,box.y0+0.02,box.width*0.9,box.height])
-
-    p1.set_ylim(np.min(synflux[:])-3, np.max(synflux[:])+3)
-
-    p1.vlines(x = em_lines*(1+z_backgal),ymin= -100,ymax= 100,colors= 'g',linestyles='dashed')
-    p1.legend(loc='upper right', bbox_to_anchor = (1.2,1), ncol = 1, prop=fontP)
-    p1.set_xlim(3500,10500)
-    plt.ylabel('Flux [$10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}]$')
-
-    p2 = plt.subplot(gs[1,:1])
-    p2.vlines(x = em_lines*(1+z_backgal),ymin= -100,ymax= 100,colors= 'g',linestyles='dashed')
-    loc_flux = flux[wave2bin((1+z_backgal)*(3727-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(3727+10),c0,c1,Nmax)]
-    p2.plot(wave[wave2bin((1+z_backgal)*(3727-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(3727+10),c0,c1,Nmax)],loc_flux,'k', label = 'OII', drawstyle='steps-mid')
-    p2.plot(wave[wave2bin((1+z_backgal)*(3727-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(3727+10),c0,c1,Nmax)],synflux[wave2bin((1+z_backgal)*(3727-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(3727+10),c0,c1,Nmax)],'r', label = 'OII', drawstyle='steps-mid')
-    if loc_flux != []:
-
-        p2.set_ylim(np.min(loc_flux)-1,np.max(loc_flux)+1)
-
-    plt.title('[OII] 3727')
-    p2.set_xlim((1+z_backgal)*(3727-10),(1+z_backgal)*(3727+10))
-    x1 = int((1+z_backgal)*3727)
-    plt.xticks([x1-15,x1,x1+15])
-    plt.ylabel('Flux [$10^{-17} erg\, s^{-1} cm^{-2}  \AA^{-1}]$')
-
-    #If Ha is below 9500 A, show it
-    if z>0.44:
-        p3 = plt.subplot(gs[1,1:4])
-    else:
-        p3 = plt.subplot(gs[1,1:3])
-    p3.vlines(x = em_lines*(1+z_backgal),ymin= -100,ymax= 100,colors= 'g',linestyles='dashed')
-    loc_flux = flux[wave2bin((1+z_backgal)*(4861-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(5007+10),c0,c1,Nmax)]
-    p3.plot(wave[wave2bin((1+z_backgal)*(4861-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(5007+10),c0,c1,Nmax)],loc_flux,'k', label = 'OIII, Hb', drawstyle='steps-mid')
-    p3.plot(wave[wave2bin((1+z_backgal)*(4861-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(5007+10),c0,c1,Nmax)],synflux[wave2bin((1+z_backgal)*(4861-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(5007+10),c0,c1,Nmax)],'r', label = 'OIII, Hb', drawstyle='steps-mid')
-    if loc_flux != []:
-
-        p3.set_ylim(np.min(loc_flux)-1,np.max(loc_flux)+1)
-
-    plt.title(r'H$\beta$,[OIII] 4959, [OIII] 5007')
-    plt.xlabel(r'Observed wavelength [$\AA$]')
-    p3.set_xlim((1+z_backgal)*(4861-10),(1+z_backgal)*(5007+10))
-    x1 = int((1+z_backgal)*4862/10.)*10
-
-    if x1<7600:
-        plt.xticks([x1,x1+50 , x1+100 ,  x1 +150 ,x1+200])
-    else:
-        plt.xticks([x1,x1+50 , x1+100 ,  x1 +150 ,x1+200, x1+ 250])
-
-
-    box = p3.get_position()
-    p3.set_position([box.x0+0.02,box.y0,box.width*0.9,box.height])
-
-    if z<0.44:
-        p4 = plt.subplot(gs[1,3:4])
-        p4.vlines(x = em_lines*(1+z_backgal),ymin= -100,ymax= 100,colors= 'g',linestyles='dashed')
-        loc_flux = flux[wave2bin((1+z_backgal)*(6562-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(6562+10),c0,c1,Nmax)]
-        p4.plot(wave[wave2bin((1+z_backgal)*(6562-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(6562+10),c0,c1,Nmax)],loc_flux,'k', label = 'Ha', drawstyle='steps-mid')
-        p4.plot(wave[wave2bin((1+z_backgal)*(6562-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(6562+10),c0,c1,Nmax)],synflux[wave2bin((1+z_backgal)*(6562-10),c0,c1,Nmax) :wave2bin((1+z_backgal)*(6562+10),c0,c1,Nmax)],'r', label = 'Ha', drawstyle='steps-mid')
-        if loc_flux != []:
-
-            p4.set_ylim(np.min(loc_flux)-1,np.max(loc_flux)+1)
-        plt.title(r'H$\alpha$')
-        p4.set_xlim((1+z_backgal)*(6562-10),(1+z_backgal)*(6562+10))
-        x1 = int((1+z_backgal)*6562)
-        if x1 < 9900:
-            plt.xticks([x1-10,x1,x1+10], [str(x1-10),str(x1),str(x1+10)])
-        else:
-            plt.xticks([x1-10,x1,x1+10], [str(x1-10),'',str(x1+10)])
-    make_sure_path_exists(topdir + savedir +'/plots/')
-    plt.savefig(topdir + savedir +'/plots/'+SDSSname(RA,DEC)+ '-' + str(plate) + '-' + str(mjd) + '-' + str(fiberid) + '-'+str(k+1) +'.png')
-
-    if show:
-        plt.show()
-    else:
-        plt.close()
-
-
